+++ swift/trainers/trainers.py
@@ -16,6 +16,8 @@ from transformers.models.auto.modeling_auto import \
     MODEL_FOR_CAUSAL_LM_MAPPING_NAMES
 from transformers.utils import is_peft_available
 
+from swift.torchacc_utils import (ta_eval_dataloader, ta_test_dataloader,
+                                  ta_train_dataloader)
 from swift.utils import use_torchacc
 from .callback import (DefaultFlowCallbackNew, PrinterCallbackNew,
                        ProgressCallbackNew)
@@ -268,32 +270,6 @@ class Seq2SeqTrainer(PushToMsHubMixin, SwiftMixin, HfSeq2SeqTrainer):
         if not use_torchacc():
             return super().get_train_dataloader()
         else:
-            # patch skip_first_batches for customized dataloader.
-            def acc_skip_first_batches(dataloader, num_batches=0):
-                from accelerate.data_loader import SkipBatchSampler
-                batch_sampler = SkipBatchSampler(
-                    dataloader._loader.batch_sampler, skip_batches=num_batches)
-                dataset = dataloader.dataset
-                dataloader_params = {
-                    'collate_fn': data_collator,
-                    'num_workers': self.args.dataloader_num_workers,
-                    'pin_memory': self.args.dataloader_pin_memory,
-                    'persistent_workers':
-                    self.args.dataloader_persistent_workers,
-                }
-
-                if not isinstance(train_dataset,
-                                  torch.utils.data.IterableDataset):
-                    dataloader_params['batch_sampler'] = batch_sampler
-                    dataloader_params['worker_init_fn'] = trainer.seed_worker
-
-                return ta.AsyncLoader(
-                    DataLoader(dataset, **dataloader_params), self.args.device)
-
-            trainer.skip_first_batches = acc_skip_first_batches
-
-            # dataloader for TorchAcc.
-            import torchacc as ta
             if trainer.is_datasets_available():
                 import datasets
 
@@ -310,22 +286,10 @@ class Seq2SeqTrainer(PushToMsHubMixin, SwiftMixin, HfSeq2SeqTrainer):
             else:
                 data_collator = self._get_collator_with_removed_columns(
                     data_collator, description='training')
-            dataloader_params = {
-                'batch_size': self._train_batch_size,
-                'collate_fn': data_collator,
-                'num_workers': self.args.dataloader_num_workers,
-                'pin_memory': self.args.dataloader_pin_memory,
-                'persistent_workers': self.args.dataloader_persistent_workers,
-            }
 
-            if not isinstance(train_dataset, torch.utils.data.IterableDataset):
-                dataloader_params['sampler'] = self._get_train_sampler()
-                dataloader_params['drop_last'] = self.args.dataloader_drop_last
-                dataloader_params['worker_init_fn'] = trainer.seed_worker
-
-            return ta.AsyncLoader(
-                DataLoader(train_dataset, **dataloader_params),
-                self.args.device)
+            return ta_train_dataloader(train_dataset, data_collator,
+                                       self._get_train_sampler(), self.args,
+                                       self._train_batch_size)
 
     def get_eval_dataloader(self, eval_dataset):
         if not use_torchacc():
@@ -349,22 +313,9 @@ class Seq2SeqTrainer(PushToMsHubMixin, SwiftMixin, HfSeq2SeqTrainer):
                 data_collator = self._get_collator_with_removed_columns(
                     data_collator, description='evaluation')
 
-            dataloader_params = {
-                'batch_size': self.args.eval_batch_size,
-                'collate_fn': data_collator,
-                'num_workers': self.args.dataloader_num_workers,
-                'pin_memory': self.args.dataloader_pin_memory,
-                'persistent_workers': self.args.dataloader_persistent_workers,
-            }
-
-            if not isinstance(eval_dataset, torch.utils.data.IterableDataset):
-                dataloader_params['sampler'] = self._get_eval_sampler(
-                    eval_dataset)
-                dataloader_params['drop_last'] = self.args.dataloader_drop_last
-
-            return ta.AsyncLoader(
-                DataLoader(eval_dataset, **dataloader_params),
-                self.args.device)
+            return ta_eval_dataloader(eval_dataset, data_collator,
+                                      self._get_eval_sampler(eval_dataset),
+                                      self.args)
 
     def get_test_dataloader(self, test_dataset):
         if not use_torchacc():
@@ -384,23 +335,9 @@ class Seq2SeqTrainer(PushToMsHubMixin, SwiftMixin, HfSeq2SeqTrainer):
                 data_collator = self._get_collator_with_removed_columns(
                     data_collator, description='test')
 
-            dataloader_params = {
-                'batch_size': self.args.eval_batch_size,
-                'collate_fn': data_collator,
-                'num_workers': self.args.dataloader_num_workers,
-                'pin_memory': self.args.dataloader_pin_memory,
-                'persistent_workers': self.args.dataloader_persistent_workers,
-            }
-
-            if not isinstance(test_dataset, torch.utils.data.IterableDataset):
-                dataloader_params['sampler'] = self._get_eval_sampler(
-                    test_dataset)
-                dataloader_params['drop_last'] = self.args.dataloader_drop_last
-
-            # We use the same batch_size as for eval.
-            return ta.AsyncLoader(
-                DataLoader(test_dataset, **dataloader_params),
-                self.args.device)
+            return ta_test_dataloader(test_dataset, data_collator,
+                                      self._get_eval_sampler(test_dataset),
+                                      self.args)
 
 
 # monkey patching
